backend/requirements.txt
# Smart IoT Predictive Maintenance - Backend
paho-mqtt>=2.1.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
google-genai>=1.0.0
joblib>=1.3.0
python-dotenv>=1.0.0
